fix: allow about:blank in the URL guard so tab-recovery works (#1)
A closed tab left agents unable to recover: the hermes browser tool reopens a
fresh tab with 'open about:blank', but validateUrl rejected any non-http(s)
scheme, so tab-create 400'd ('Blocked URL scheme: about:') and the agent could
never get a working surface back. Observed live — an agent told to stop had its
tab closed, then spun retrying a dead tab / failing to reopen.

about:blank is the canonical empty page (a freshly created page already sits on
it, and the server itself does page.goto('about:blank') elsewhere); the scheme
guard is meant to block file:/chrome:/data:, not the inert blank page.

Moved validateUrl into lib/request-utils.js (where the sibling error classifier
already lives) so it is unit-testable, added the about:blank allowance, removed
the now-dead ALLOWED_URL_SCHEMES const from server.js. New unit test covers
about:blank/http/https allowed, file/chrome/data/about:config blocked, garbage
rejected. Verified all cases green.
