+++ lib/request-utils.js
@@ -1,6 +1,31 @@
 // HTTP request classification helpers -- kept separate from metrics.js
 // Separated from server.js to keep HTTP method classification in its own module.
 
+const ALLOWED_URL_SCHEMES = ['http:', 'https:'];
+
+/**
+ * Validate a navigation URL. Returns null if allowed, or an error string.
+ *
+ * `about:blank` is explicitly allowed: it is the canonical empty page (a freshly
+ * created tab already sits on it), and clients legitimately request it when
+ * opening a blank tab to recover after a tab closes. Rejecting it deadlocked
+ * agents whose tab was closed — their reopen ("open about:blank") 400'd, so they
+ * could never get a working surface back. The scheme guard exists to block
+ * file:/chrome:/data: etc., not the inert blank page.
+ */
+export function validateUrl(url) {
+  if (url === 'about:blank') return null;
+  try {
+    const parsed = new URL(url);
+    if (!ALLOWED_URL_SCHEMES.includes(parsed.protocol)) {
+      return `Blocked URL scheme: ${parsed.protocol} (only http/https allowed)`;
+    }
+    return null;
+  } catch {
+    return `Invalid URL: ${url}`;
+  }
+}
+
 /**
  * Derive a short action name from an Express request for metrics labeling.
  */
